+++ lib/resources/emojis.dart
@@ -1,0 +1,298 @@
+// Auto-generated file. Do not edit.
+
+// ignore_for_file: non_constant_identifier_names
+
+class Emoji {
+  String name;
+  String emoji;
+
+  Emoji(this.name, this.emoji);
+}
+
+Emoji heroxp = Emoji("AccountResource:heroxp", "<:heroxp:849067890824708096>");
+
+Emoji reagent_alteration_upgrade_vr = Emoji(
+    "AccountResource:reagent_alteration_upgrade_vr",
+    "<:epicperkup:849068923315683368>");
+
+Emoji reagent_alteration_upgrade_sr = Emoji(
+    "AccountResource:reagent_alteration_upgrade_sr",
+    "<:legendaryperkup:849068914071830539>");
+
+Emoji legendaryperkup =
+    Emoji("Legendary PERK-UP!", "<:legendaryperkup:849068914071830539>");
+
+Emoji reagent_alteration_upgrade_r = Emoji(
+    "AccountResource:reagent_alteration_upgrade_r",
+    "<:rareperkup:849068924569255936>");
+
+Emoji reagent_alteration_upgrade_uc = Emoji(
+    "AccountResource:reagent_alteration_upgrade_uc",
+    "<:uncommonperkup:849068926167154718>");
+
+Emoji reagent_alteration_generic = Emoji(
+    "AccountResource:reagent_alteration_generic",
+    "<:reperk:849068927778422794>");
+
+Emoji reagent_alteration_gameplay_generic = Emoji(
+    "AccountResource:reagent_alteration_gameplay_generic",
+    "<:corereperk:849068917603041280>");
+
+Emoji schematicxp =
+    Emoji("AccountResource:schematicxp", "<:schematicxp:849067891692666960>");
+
+Emoji personnelxp =
+    Emoji("AccountResource:personnelxp", "<:survivorxp:849067892934574100>");
+
+Emoji phoenixxp =
+    Emoji("AccountResource:phoenixxp", "<:venturexp:849067894624223242>");
+
+Emoji reagent_alteration_ele_fire = Emoji(
+    "AccountResource:reagent_alteration_ele_fire",
+    "<:fireup:849068930252931082>");
+
+Emoji reagent_alteration_ele_water = Emoji(
+    "AccountResource:reagent_alteration_ele_water",
+    "<:frostup:849068931947036682>");
+
+Emoji reagent_alteration_ele_nature = Emoji(
+    "AccountResource:reagent_alteration_ele_nature",
+    "<:ampup:849068933393678336>");
+
+Emoji reagent_c_t01 = Emoji(
+    "AccountResource:reagent_c_t01", "<:puredropofrain:849067869300195369>");
+
+Emoji reagent_c_t02 = Emoji("AccountResource:reagent_c_t02",
+    "<:lightninginabottle:849067871086837790>");
+
+Emoji reagent_c_t03 = Emoji(
+    "AccountResource:reagent_c_t03", "<:eyeofthestorm:849067872122961981>");
+
+Emoji reagent_c_t04 =
+    Emoji("AccountResource:reagent_c_t04", "<:stormshard:849067874136096818>");
+
+Emoji reagent_people = Emoji(
+    "AccountResource:reagent_people", "<:trainingmanual:849062319794815026>");
+
+Emoji reagent_weapons = Emoji(
+    "AccountResource:reagent_weapons", "<:weapondesigns:849062321325867048>");
+
+Emoji reagent_traps =
+    Emoji("AccountResource:reagent_traps", "<:trapdesigns:849062322953125978>");
+
+Emoji reagent_promotion_heroes = Emoji(
+    "AccountResource:reagent_promotion_heroes",
+    "<:herosupercharger:849067856863690792>");
+
+Emoji herosupercharger =
+    Emoji("Hero Supercharger", "<:herosupercharger:849067856863690792>");
+
+Emoji reagent_promotion_survivors = Emoji(
+    "AccountResource:reagent_promotion_survivors",
+    "<:survivorsupercharger:849067858486231050>");
+
+Emoji survivorsupercharger = Emoji(
+    "Survivor Supercharger", "<:survivorsupercharger:849067858486231050>");
+
+Emoji reagent_promotion_weapons = Emoji(
+    "AccountResource:reagent_promotion_weapons",
+    "<:weaponsupercharger:849067860344438834>");
+
+Emoji weaponsupercharger =
+    Emoji("Weapon Supercharger", "<:weaponsupercharger:849067860344438834>");
+
+Emoji reagent_promotion_traps = Emoji("AccountResource:reagent_promotion_traps",
+    "<:trapsupercharger:849067862232137819>");
+
+Emoji trapsupercharger =
+    Emoji("Trap Supercharger", "<:trapsupercharger:849067862232137819>");
+
+Emoji reagent_evolverarity_r = Emoji(
+    "AccountResource:reagent_evolverarity_r", "<:rareflux:849067878506299402>");
+
+Emoji reagent_evolverarity_vr = Emoji("AccountResource:reagent_evolverarity_vr",
+    "<:epicflux:849067875879747644>");
+
+Emoji reagent_evolverarity_sr = Emoji("AccountResource:reagent_evolverarity_sr",
+    "<:legendaryflux:849067863826235442>");
+
+Emoji voucher_item_buyback = Emoji("AccountResource:voucher_item_buyback",
+    "<:weaponresearchvoucher:849062314317578272>");
+
+Emoji eventcurrency_scaling = Emoji(
+    "AccountResource:eventcurrency_scaling", "<:gold:849062318306099231>");
+
+Emoji voucher_cardpack_bronze = Emoji("AccountResource:voucher_cardpack_bronze",
+    "<:upgradellama:849067880452194334>");
+
+Emoji voucher_herobuyback = Emoji("AccountResource:voucher_herobuyback",
+    "<:herorecruitmentvoucher:849062316028985344>");
+
+Emoji voucher_basicpack = Emoji("AccountResource:voucher_basicpack",
+    "<:minirewardllama:849067895682367498>");
+
+Emoji voucher_cardpack_jackpot = Emoji(
+    "AccountResource:voucher_cardpack_jackpot", "<:gold:839621889114112040>");
+
+Emoji currency_xrayllama = Emoji(
+    "AccountResource:currency_xrayllama", "<:xraytickets:849064139648663584>");
+
+Emoji smallxpboost = Emoji(
+    "ConsumableAccountItem:smallxpboost", "<:xpboost:849067883278368818>");
+
+Emoji smallxpboost_gift = Emoji("ConsumableAccountItem:smallxpboost_gift",
+    "<:teammatexpboost:849067884839960576>");
+
+Emoji fortitude = Emoji("fortitude", "<:fortitude:803327857056284722>");
+
+Emoji resistance = Emoji("resistance", "<:resistance:803430864431874098>");
+
+Emoji offense = Emoji("offense", "<:offense:803430849105887273>");
+
+Emoji tech = Emoji("tech", "<:technology:803430881724596255>");
+
+Emoji technology = Emoji("technology", "<:technology:803430881724596255>");
+
+Emoji talentedbuilder =
+    Emoji("Talented Builder", "<:build:803429012113457162>");
+
+Emoji gognome = Emoji("Go Gnome!", "<:gnome:803429049673318481>");
+
+Emoji lootlegend = Emoji("Loot Legend", "<:treasure:803429115378008094>");
+
+Emoji guardianangel = Emoji("Guardian Angel", "<:survivor:803429143236182037>");
+
+Emoji playswellwithothers =
+    Emoji("Plays Well with Others", "<:pwo:803429162924769311>");
+
+Emoji unspeakablehorrors =
+    Emoji("Unspeakable Horrors", "<:horror:803429219727048735>");
+
+Emoji worldexplorer = Emoji("World Explorer", "<:explorer:803429243996340235>");
+
+Emoji research = Emoji("research", "<:research:803433855393988608>");
+
+Emoji vbucks = Emoji("vbucks", "<:vbucks:803432012358615051>");
+
+Emoji currency_mtxswap =
+    Emoji("AccountResource:currency_mtxswap", "<:vbucks:803432012358615051>");
+
+Emoji stormkingsscourge =
+    Emoji("stw:stormkings-scourge", "<:stormkingsscourge:850673955017719829>");
+
+Emoji stormkingsfury =
+    Emoji("stw:stormkings-fury", "<:stormkingsfury:850673849535299585>");
+
+Emoji stormkingsravager =
+    Emoji("stw:stormkings-ravager", "<:stormkingsravager:850674233616891925>");
+
+Emoji stormkingswrath =
+    Emoji("stw:stormkings-wrath", "<:stormkingswrath:850674078167334943>");
+
+Emoji stormkingsonslaught = Emoji(
+    "stw:stormkings-onslaught", "<:stormkingsonslaught:850674313655877642>");
+
+Emoji seasonxp = Emoji("Season XP", "<:xp:836010555868708924>");
+
+Emoji tick = Emoji("tick", "<:tick:901487022005436477>");
+
+Emoji cross = Emoji("cross", "<:no:901487039474696232>");
+
+Emoji loading = Emoji("loading", "<a:loading:804391868376416306>");
+
+Emoji common = Emoji("common", "<:common:901830595259293706>");
+
+Emoji uncommon = Emoji("uncommon", "<:uncommon:901830611503841301>");
+
+Emoji rare = Emoji("rare", "<:rare:901830628209737738>");
+
+Emoji epic = Emoji("epic", "<:epic:901830656013770853>");
+
+Emoji legendary = Emoji("legendary", "<:legendary:901830667631992893>");
+
+Emoji mythic = Emoji("mythic", "<:mythic:901830677048221706>");
+
+Emoji star = Emoji("star", "<:star:907276087174778900>");
+
+Emoji paidstar = Emoji("paidStar", "<:bp:914863340923260928>");
+
+List<Emoji> emojis = [
+  heroxp,
+  reagent_alteration_upgrade_vr,
+  reagent_alteration_upgrade_sr,
+  legendaryperkup,
+  reagent_alteration_upgrade_r,
+  reagent_alteration_upgrade_uc,
+  reagent_alteration_generic,
+  reagent_alteration_gameplay_generic,
+  schematicxp,
+  personnelxp,
+  phoenixxp,
+  reagent_alteration_ele_fire,
+  reagent_alteration_ele_water,
+  reagent_alteration_ele_nature,
+  reagent_c_t01,
+  reagent_c_t02,
+  reagent_c_t03,
+  reagent_c_t04,
+  reagent_people,
+  reagent_weapons,
+  reagent_traps,
+  reagent_promotion_heroes,
+  herosupercharger,
+  reagent_promotion_survivors,
+  survivorsupercharger,
+  reagent_promotion_weapons,
+  weaponsupercharger,
+  reagent_promotion_traps,
+  trapsupercharger,
+  reagent_evolverarity_r,
+  reagent_evolverarity_vr,
+  reagent_evolverarity_sr,
+  voucher_item_buyback,
+  eventcurrency_scaling,
+  voucher_cardpack_bronze,
+  voucher_herobuyback,
+  voucher_basicpack,
+  voucher_cardpack_jackpot,
+  currency_xrayllama,
+  smallxpboost,
+  smallxpboost_gift,
+  fortitude,
+  resistance,
+  offense,
+  tech,
+  technology,
+  talentedbuilder,
+  gognome,
+  lootlegend,
+  guardianangel,
+  playswellwithothers,
+  unspeakablehorrors,
+  worldexplorer,
+  research,
+  vbucks,
+  currency_mtxswap,
+  stormkingsscourge,
+  stormkingsfury,
+  stormkingsravager,
+  stormkingswrath,
+  stormkingsonslaught,
+  stormkingsfury,
+  stormkingsscourge,
+  stormkingswrath,
+  stormkingsravager,
+  stormkingsonslaught,
+  seasonxp,
+  tick,
+  cross,
+  loading,
+  common,
+  uncommon,
+  rare,
+  epic,
+  legendary,
+  mythic,
+  star,
+  paidstar,
+];